Add -w/--wslwindows option
This also adds a specific for WSL2.

This option executes the operation on the windows host, if run with a
WSL2 Kernel.

There currently are multiple things hard-coded. The windows installation
needs to be in the C drive and it of course needs to be mounted inside
of WSL. WSL needs to have the windows interoperability that comes by
default to execute the cmd.exe.

This also reworks the way the command is built to be way cleaner and
easier to work with in the future